Move signals out of models.py per django docs recommendation
