+++ arches_for_science/models.py
@@ -1,10 +1,7 @@
 import uuid
-
-from arches.app.models.models import IIIFManifest, TileModel, FunctionXGraph
+from arches.app.models.models import TileModel, FunctionXGraph
 from django.db import models
 from django.db.models import JSONField
-from django.db.models.signals import post_delete, post_save
-from django.dispatch import receiver
 import pgtrigger
 
 from .trigger_functions import CALCULATE_MULTICARD_PRIMARY_DESCRIPTOR_SINGLE, CALCULATE_MULTICARD_PRIMARY_DESCRIPTOR_ALL
@@ -82,17 +79,3 @@ class ManifestXCanvas(models.Model):
     class Meta:
         managed = True
         db_table = "manifest_x_canvas"
-
-
-@receiver(post_save, sender=IIIFManifest)
-def create_digital_resources(sender, instance, created, **kwargs):
-    from arches_for_science.utils.digital_resource_for_manifest import digital_resources_for_manifest, digital_resources_for_canvases
-
-    digital_resources_for_manifest(instance, created)
-    digital_resources_for_canvases(instance)
-
-
-@receiver(post_delete, sender=IIIFManifest)
-def delete_manifest_x_canvas(sender, instance, **kwargs):
-    ManifestXCanvas.objects.filter(manifest=instance.manifest["@id"]).delete()
-    ManifestXDigitalResource.objects.filter(manifest=instance.manifest["@id"]).delete()
